requirements.txt
streamlit>=1.32.0
requests>=2.31.0
pandas>=2.0.0
speedtest-cli>=2.1.3
